+++ src/common/styles/realtimeDefaultStyle.ts
@@ -62,7 +62,7 @@ export const getDelayTextCanvas = (
       Math.ceil(fontSize + 8 * pixelRatio),
     );
     if (canvas) {
-      const ctx = canvas.getContext('2d');
+      const ctx = canvas.getContext('2d') as CanvasRenderingContext2D;
       if (!ctx) {
         return null;
       }
@@ -146,6 +146,7 @@ export const getTextCanvas = (
   strokeColor: string,
   hasStroke: boolean,
   pixelRatio: number,
+  getTextFont: (fontSize: number, text?: string) => string,
 ) => {
   const key = `${text}, ${origin}, ${textSize}, ${fillColor},${strokeColor}, ${hasStroke}, ${pixelRatio}`;
   if (!cacheText[key]) {
@@ -161,7 +162,7 @@ export const getTextCanvas = (
         ctx.save();
         ctx.textBaseline = 'middle';
         ctx.textAlign = 'center';
-        ctx.font = `bold ${textSize + 2}px Arial`;
+        ctx.font = getTextFont(textSize + 2, text);
         ctx.strokeStyle = strokeColor;
         ctx.strokeText(text, origin, origin);
         ctx.restore();
@@ -171,7 +172,7 @@ export const getTextCanvas = (
       ctx.textBaseline = 'middle';
       ctx.textAlign = 'center';
       ctx.fillStyle = fillColor;
-      ctx.font = `bold ${textSize}px Arial`;
+      ctx.font = getTextFont(textSize, text);
       ctx.strokeStyle = strokeColor;
       ctx.strokeText(text, origin, origin);
       ctx.fillText(text, origin, origin);
@@ -208,8 +209,10 @@ const realtimeDefaultStyle: RealtimeStyleFunction = (
     getBgColor = () => '#000',
     getDelayColor = () => '#000',
     getDelayText = () => null,
+    getDelayFont = (fontSize: number) => `bold ${fontSize}px arial, sans-serif`,
+    getTextFont = (fontSize: number) => `bold ${fontSize}px arial, sans-serif`,
     getTextColor = () => '#000',
-    getTextSize = () => 0,
+    getTextSize = () => 14,
     getMaxRadiusForText = () => 10,
     getMaxRadiusForStrokeAndDelay = () => 7,
   } = options;
@@ -324,7 +327,7 @@ const realtimeDefaultStyle: RealtimeStyleFunction = (
         delayText = getDelayTextCanvas(
           text,
           fontSize,
-          `bold ${fontSize}px arial, sans-serif`,
+          getDelayFont(fontSize, text),
           getDelayColor(delay, cancelled, true),
           delayOutlineColor,
           pixelRatio,
@@ -362,7 +365,7 @@ const realtimeDefaultStyle: RealtimeStyleFunction = (
     const height = size;
     const canvas = createCanvas(width, height);
     if (canvas) {
-      const ctx = canvas.getContext('2d');
+      const ctx = canvas.getContext('2d') as CanvasRenderingContext2D;
       if (!ctx) {
         return null;
       }
@@ -382,7 +385,13 @@ const realtimeDefaultStyle: RealtimeStyleFunction = (
       let circleText = null;
       if (isDisplayText) {
         const fontSize2 = Math.max(radius, 10);
-        const textSize = getTextSize(ctx, markerSize, name, fontSize2);
+        const textSize = getTextSize(
+          ctx,
+          markerSize,
+          name,
+          fontSize2,
+          getTextFont,
+        );
         const textColor2 = !useDelayStyle
           ? textColor || getTextColor(type)
           : '#000000';
@@ -399,6 +408,7 @@ const realtimeDefaultStyle: RealtimeStyleFunction = (
           circleFillColor,
           hasStroke2,
           pixelRatio,
+          getTextFont,
         );
       }