+++ CHANGELOG.md
@@ -2,6 +2,39 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [2.3.5](https://github.com/geops/mobility-toolbox-js/compare/v2.3.4...v2.3.5) (2023-11-17)
+
+
+### Bug Fixes
+
+* simplify setBbox function ([8f99991](https://github.com/geops/mobility-toolbox-js/commit/8f99991df3f99c64b3776cd8c18dff0820c4a638))
+* use floored value for the zoom ([7961de0](https://github.com/geops/mobility-toolbox-js/commit/7961de0d3e660d67658205aef5e6578ea79b440b))
+
+### [2.3.4](https://github.com/geops/mobility-toolbox-js/compare/v2.3.3...v2.3.4) (2023-11-17)
+
+
+### Bug Fixes
+
+* do not set generalizationLevelByZoom by default ([69b59ba](https://github.com/geops/mobility-toolbox-js/commit/69b59ba204529eb90cf1bd99a5bedd95705f821f))
+* simplify getMotsByZoom behavior ([1fc259f](https://github.com/geops/mobility-toolbox-js/commit/1fc259f29bf3da276a158aa2c063e7609ca6a6f9))
+* simplify getMotsByZoom behavior ([87e52ec](https://github.com/geops/mobility-toolbox-js/commit/87e52ec26e9d7204b563a39602fa3c0aa1a90530))
+
+### [2.3.3](https://github.com/geops/mobility-toolbox-js/compare/v2.3.2...v2.3.3) (2023-11-17)
+
+
+### Bug Fixes
+
+* add text parameter to getXXXFont functions ([83cec9b](https://github.com/geops/mobility-toolbox-js/commit/83cec9b3b9a58aad76a3944257cf83dc80710b89))
+* send floor/ceil values to the websocket ([50b8b0a](https://github.com/geops/mobility-toolbox-js/commit/50b8b0aaad28f9cc63ab1c7f460d2662262bcfe6))
+* update full trajectory layer when interacting and animating ([97a9ad8](https://github.com/geops/mobility-toolbox-js/commit/97a9ad86d03d4ca8f760df3269840537746cd403))
+
+### [2.3.2](https://github.com/geops/mobility-toolbox-js/compare/v2.3.1...v2.3.2) (2023-11-14)
+
+
+### Bug Fixes
+
+* add getXXXFont function to realtime style options ([310be23](https://github.com/geops/mobility-toolbox-js/commit/310be2350e96eb815d080f1a00eaafcb8d79e1e6))
+
 ### [2.3.1](https://github.com/geops/mobility-toolbox-js/compare/v2.3.0...v2.3.1) (2023-11-03)